[net.common.c] fixed issue in the 'get_exists_functions' function.
Count of function was recently changed, however in function used own
counter, so reserver for 32 functions for now.

[net.xml] recently added function now in the 'expected_functions' list.

[tests.xml] fixed issue with not found data for the
'TestNetModuleViaBuildFile' test.

[tests_net.module.cpp] added call of SetUp functions from base classes.
Recently changed hierarchy do not handle that project was not created
after that change.

Checking of function return was change from comparing of return code to
the comparing is result Success.